+++ src/test/icalcomponent_fuzz2.c
@@ -53,11 +53,11 @@ int main(int argc, char *argv[])
         fclose(fp);
         assert(0);
     }
-    filesize = sbuf.st_size; //to make fortify compile happy
+    filesize = (size_t)sbuf.st_size; //to make fortify compile happy
     data = malloc(filesize + 1);
     memset(data, 0, filesize + 1);
 
-    r = read(fd, data, filesize);
+    r = (int)read(fd, data, filesize);
     /* cppcheck-suppress doubleFree */
     fclose(fp);
